+++ classic_adversarial_autoencoder.py
@@ -289,7 +289,7 @@ def parse_arguments(argv):
     parser.add_argument('--epochs', type=int, help='number of iterations to train', default=100)
     parser.add_argument('--ae_it', type=int,
                         help='number of epochs to pretrain the autoencoder', default=0)
-    parser.add_argument('--train_prop', type=int, help='Proportion of train set', default=0.8)
+    parser.add_argument('--train_prop', type=float, help='Proportion of train set', default=0.8)
     return parser.parse_args(argv)
